Moving internal method tests to a new class
